+++ Models/GW1005/Response/ItakushaJoho.cs
@@ -2,7 +2,7 @@ namespace WebAPIJsonDataMaker.Models.GW1005.Response
 {
     public class ItakushaJoho
     {
-        public int itakushaCode { get; set; }
+        public long itakushaCode { get; set; }
         public int shikinkessaiTemban { get; set; }
         public int shikinkessaiKamoku { get; set; }
         public int shikinkessaiKozabango { get; set; }
