+++ Models/GW1006/JiMeisaiKey.cs
@@ -2,6 +2,6 @@ namespace WebAPIJsonDataMaker.Models.GW1006
 {
     public class JiMeisaiKey
     {
-        public int tsugiItakushaCode{get;set;}
+        public long tsugiItakushaCode{get;set;}
     }
 }
